Remove `Acknowledgement` class in favor of just setting the properties in the `OfflineResponse` directly.
