--- test/Intacct/Xml/Response/AcknowledgementTest.php
@@ -1,92 +1,0 @@
-<?php
-/**
- * Copyright 2017 Intacct Corporation.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License"). You may not
- *  use this file except in compliance with the License. You may obtain a copy
- *  of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * or in the "LICENSE" file accompanying this file. This file is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- *
- */
-
-namespace Intacct\Xml\Response;
-
-use Intacct\Xml\OfflineResponse;
-
-/**
- * @coversDefaultClass \Intacct\Xml\Response\Acknowledgement
- */
-class AcknowledgementTest extends \PHPUnit\Framework\TestCase
-{
-
-    /**
-     * @var Acknowledgement
-     */
-    protected $object;
-
-    /**
-     * Sets up the fixture, for example, opens a network connection.
-     * This method is called before a test is executed.
-     */
-    protected function setUp()
-    {
-        $xml = <<<EOF
-<?xml version="1.0" encoding="UTF-8"?>
-<response>
-      <acknowledgement>
-            <status>success</status>
-      </acknowledgement>
-      <control>
-            <status>success</status>
-            <senderid>testsenderid</senderid>
-            <controlid>ControlIdHere</controlid>
-            <uniqueid>false</uniqueid>
-            <dtdversion>3.0</dtdversion>
-      </control>
-</response>
-EOF;
-        $response = new OfflineResponse($xml);
-        $this->object = $response->getAcknowledgement();
-    }
-
-    /**
-     * Tears down the fixture, for example, closes a network connection.
-     * This method is called after a test is executed.
-     */
-    protected function tearDown()
-    {
-    }
-
-    public function testGetStatus()
-    {
-        $this->assertEquals('success', $this->object->getStatus());
-    }
-    
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage Acknowledgement block is missing status element
-     */
-    public function testMissingStatusElement()
-    {
-        $xml = <<<EOF
-<?xml version="1.0" encoding="UTF-8"?>
-<response>
-      <acknowledgement/>
-      <control>
-            <status>success</status>
-            <senderid>testsenderid</senderid>
-            <controlid>ControlIdHere</controlid>
-            <uniqueid>false</uniqueid>
-            <dtdversion>3.0</dtdversion>
-      </control>
-</response>
-EOF;
-        new OfflineResponse($xml);
-    }
-}
